+++ src/main/java/cloud/metaapi/sdk/clients/meta_api/models/MetatraderPosition.java
@@ -34,7 +34,7 @@ public class MetatraderPosition {
     /**
      * Position magic number, identifies the EA which opened the position
      */
-    public int magic;
+    public long magic;
     /**
      * Time position was opened at
      */
